# Make FakeValuesService thread safe