--- a/src/main/java/net/datafaker/internal/helper/COWMap.java
+++ b/src/main/java/net/datafaker/internal/helper/COWMap.java
@@ -87,8 +87,7 @@ public class COWMap<K, V> implements Map<K, V> {
 
     public <K2, V2> void updateNestedValue(K key, Supplier<V> valueSupplier, K2 key2, V2 value) {
         map.putIfAbsent(key, valueSupplier.get());
-        // It is assumed that nested could be only StampedLockMap
-        // otherwise there is no guarantee for thread safe
+        // It is assumed that nested could be only Map
         ((Map<K2, V2>)map.get(key)).put(key2, value);
     }
 }