The PR assumes that most of the time there is no writes to the maps and only reads.
~In this case there could be used optimistic locks most of the time and cheap `StampedLock` 's~

~based on measurements there is overhead only about 10% for some of methods, for others there is no noticeable difference~
e.g.
`Datafaker_KotlinFakerBenchmark.kotlinFakerTest`
before PR
```
Benchmark                                       Mode  Cnt    Score   Error  Units
Datafaker_KotlinFakerBenchmark.kotlinFakerTest  avgt   10  390.813 ± 2.512  ms/op
```
after (10% slower)
```
Benchmark                                       Mode  Cnt    Score    Error  Units
Datafaker_KotlinFakerBenchmark.kotlinFakerTest  avgt   10  431.599 ± 13.887  ms/op
```
`DatafakerSimpleMethods.firstname`
before
```
Benchmark                              Mode  Cnt      Score     Error   Units
DatafakerSimpleMethods.firstname      thrpt   10   6016.154 ± 247.178  ops/ms
```
after (about 10% slower)
```
Benchmark                              Mode  Cnt      Score      Error   Units
DatafakerSimpleMethods.firstname      thrpt   10   5330.902 ±   67.988  ops/ms
```
`DatafakerSimpleMethods.fullname`
before
```
Benchmark                              Mode  Cnt      Score     Error   Units
DatafakerSimpleMethods.fullname       thrpt   10   2524.751 ±  26.856  ops/ms
```
after (about 10% slower)
```
Benchmark                              Mode  Cnt      Score      Error   Units
DatafakerSimpleMethods.fullname       thrpt   10   2235.404 ±   67.011  ops/ms
```

UPD: a new approach passing added test is with `COWMap`