@rwong-gw 2023-04-20
For the past couple days, I have been using CI to run my Gatling tests every half hour with this version, and everything looks good, with no hanging or deadlocks.

One suggestion: `StampedLockMap` does not implement several methods in Java's `Map` API, and just throws an `UnsupportedOperationException` for them. This causes it to interact poorly with debuggers (see attached image from IntelliJ). You should implement at least the `size` and `entrySet` methods to get it to work nicely with IntelliJ, and it probably wouldn't hurt to implement at least the remaining non-mutating operations.

![StampedLockMap_UOE](https://user-images.githubusercontent.com/130408730/233494393-8b6c6b3c-2ac9-434a-b6f7-657b261c05e9.png)