@snuyanzin 2023-04-21
makes sense, added implementation