@rwong-gw 2023-04-21
Thanks, I can now drill down into StampedLockMaps in IntelliJ's debugger:
![StampedLockMap_ro](https://user-images.githubusercontent.com/130408730/233711925-04c4a8d1-a4b4-4ccd-a55c-713d4dfc3904.png)

BTW, you might want to add `@Override` annotations to the methods `putIfAbsent`, `computeIfAbsent`, and `merge`.