@snuyanzin 2023-04-23
Funny thing is that I finally adapted the test from #759 attached by @rwong-gw 
And once I decreased sleep time it starts failing with `IllegalMonitorStateException` since it can not unlock locks ...

At the same time another option I had in mind was CopyOnWriteMaps (`COWMap`) which for some reason is not provided by java core. 

So I replaced it with `COWMap` and the problem has gone.
Another good thing that it is a bit faster than with stamplockmaps (almost same speed as with unsafe option).
Probably at the beginning it could consume a bit more memory.

@kingthorin so at least one test is added
@rwong-gw could you please once again double check your case if it is still ok? I'm asking since as mentioned above I replaced it with `COWMap` because of `IllegalMonitorStateException`