@rwong-gw 2023-04-24
> @rwong-gw could you please once again double check your case if it is still ok? I'm asking since as mentioned above I replaced it with `COWMap` because of `IllegalMonitorStateException`

@snuyanzin I'll give it a try.